Travis CI: install all project dependencies
This is required for Pylint to be able to detect missing imports. We use
caching capabilities to improve the duration of the test runs.

Furthermore we fix an issue that arises when using the most recent
version of RPi.GPIO. The module would not be unloaded properly, causing
test issues because we are not running on a Raspberry Pi device. We also
address a wrong import order message that was not reported before.